+++ .gitignore
@@ -1,4 +1,5 @@
 .DS_Store
+dist
 node_modules
 package-lock.json
 release/docs
