+++ package.json
@@ -1,6 +1,9 @@
 {
   "name": "root",
   "devDependencies": {
-    "lerna": "^4.0.0"
+    "esbuild": "^0.17.18",
+    "lerna": "^4.0.0",
+    "npm-run-all": "^4.1.5",
+    "typescript": "^5.0.4"
   }
 }
